moved into EntityDefaultTest
